feat(core): add --skip-dependencies CLI option
Added a `--skip-dependencies` CLI option for the `deploy` and `test`
commands.

When used with the `deploy` command, the specified services are
deployed, but any additional services or tasks that they depend on
are skipped.

This can be useful e.g. when your stack has already been deployed,
and you want to deploy a subset of services in dev mode
without redeploying any service dependencies that may have changed
since you last deployed.

Similarly, for the `test` command, this option skips deploying/running
any runtime dependencies of the requested test suites (not recommended
for CI, but can be useful during inner-loop development in certain
scenarios).