fix(core): show error indicator in watch footer
When a task fails during a watch-mode command's execution, the footer
now indicates that one or more tasks failed (instead of saying "Waiting
for code changes...").

This has been asked for by several users, who felt that indicating an
error in the footer makes it clearer that something went wrong as the
tasks were being processed.